v6.32
Range constructors/inserts/assigns updated so that the slower form that is required to match sentinels, will not match iterator types which can be converted to one another eg. const_iterator and iterator. This means that, for example, a call to range-assign using colony's const_iterator and iterator will be able to use colony's faster ADL distance() overload rather than just doing a straight for-loop to determine length of range (necessary with sentinels, as there is no overload for distance() for sentinels).